Change readme ant task slightly
I'm not entirely sure why, but running this ant task as it was didn't behave correctly for me. It worked fine on a Mac, but didn't work on my Linux Jenkins box. I have a feeling it's something to do with my Linux Jenkins box having spaces in the file path.

This works fine though, and doesn't make any major differences I don't think.