Travis: test against nightly (= PHP 8)
Adding a build against `nightly` (PHP 8) which, for now, is allowed to fail.

At this time, there is one unit test which fails `ParallelLint.lint.phpt method=testDeprecated`. This will need to be investigated over time. No urgency yet as it may be related to the Nette testing framework dependency and also PHP 8 is still in flux.