ParallelLintLintTest::testDeprecated(): fix the test
This test is expected to throw a deprecation warning for the PHP4-style class constructor (method named the same as the class).

In PHP 8, the deprecation warning is no longer thrown and PHP4-style constructor method are now treated as _normal_ methods.

This fixes the test to have the correct expectations.